requirements.txt
Django==3.0.8
django-crispy-forms==1.8.1
channels==3.0.2
daphne==3.0.1
channels-redis==3.2.0
whitenoise==5.2.0
dj-database-url==0.5.0
djangorestframework==3.12.2